refactor peripheral /(uten testing)
