--- battery-sensor/nrf/promicro_nrf52840/peripheral/Kconfig.sysbuild
@@ -1,16 +1,0 @@
-# Copyright 2023-2024 Nordic Semiconductor ASA
-# SPDX-License-Identifier: Apache-2.0
-
-# Er denne nødvendig?
-source "share/sysbuild/Kconfig"
-
-config NET_CORE_BOARD
-	string
-	default "nrf5340dk/nrf5340/cpunet" if $(BOARD) = "nrf5340dk"
-	default "nrf5340_audio_dk/nrf5340/cpunet" if $(BOARD) = "nrf5340_audio_dk"
-	default "nrf5340bsim/nrf5340/cpunet" if $(BOARD_TARGET_STRING) = "NRF5340BSIM_NRF5340_CPUAPP"
-
-config NET_CORE_IMAGE_HCI_IPC
-	bool "HCI IPC image on network core"
-	default y
-	depends on NET_CORE_BOARD != ""
