+++ battery-sensor/nrf/promicro_nrf52840/peripheral/prj.conf
@@ -1,18 +1,9 @@
 CONFIG_BT=y
 CONFIG_LOG=y
 
-# pairing... i dont think we use it? just leave it on for now
-CONFIG_BT_SMP=n
-
 CONFIG_BT_PERIPHERAL=y
 
-# maybe DIS back
-CONFIG_BT_DIS=n
-CONFIG_BT_DIS_PNP=n
-
-CONFIG_BT_BAS=n
-CONFIG_BT_HRS=n
-CONFIG_BT_DEVICE_NAME="Kystlaget Sensor BAT-x"
+CONFIG_BT_DEVICE_NAME="Kystlaget Peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=833
 CONFIG_ADC=y
 
@@ -25,5 +16,3 @@ CONFIG_CBPRINTF_FP_SUPPORT=y
 
 # mpu
 CONFIG_MPU6050_TRIGGER_NONE=y
-# CONFIG_SHELL=y
-# CONFIG_I2C_SHELL=y
