+++ battery-sensor/nrf/promicro_nrf52840/peripheral/src/ble.h
@@ -1,0 +1,8 @@
+#ifdef BLE_H
+#define BLE_H
+
+int ble_init(void);
+
+int ble_advertise(void);
+
+#endif // BLE_H
