+++ battery-sensor/nrf/promicro_nrf52840/peripheral/src/led.h
@@ -1,0 +1,8 @@
+#ifdef LED_H
+#define LED_H
+
+void led_init(void);
+
+void led_set_connection_state(bool state);
+
+#endif // LED_H
