+++ battery-sensor/nrf/promicro_nrf52840/peripheral/src/sensor.h
@@ -1,0 +1,9 @@
+#ifndef SENSOR_H
+#define SENSOR_H
+
+#include <stddef.h>
+
+// Comment
+int sensor_init(void);
+
+#endif // SENSOR_H
