+++ battery-sensor/nrf/promicro_nrf52840/peripheral/src/sensors/uuid.h
@@ -1,0 +1,44 @@
+#ifndef SENSORS_UUID_H
+#define SENSORS_UUID_H
+
+#include <zephyr/kernel.h>
+
+#define BT_UUID_ID          BT_UUID_DECLARE_16(0x2c05)
+uint8_t custom_uuid[];
+#ifdef CONFIG_USE_ADC
+#define BT_UUID_ADC     BT_UUID_DECLARE_16(0x2c06)
+custom_uuid[] = {
+    0x00, 0x00, 0x2C, 0x06,
+    0x00, 0x00,
+    0x10, 0x00,
+    0x80, 0x00,
+    0x00, 0x80,
+    0x5F, 0x9B, 0x34, 0xFB,
+};
+#endif // CONFIG_USE_ADC
+
+#ifdef CONFIG_USE_BME280
+#define BT_UUID_TEMPERATURE BT_UUID_DECLARE_16(0x2c07)
+custom_uuid[] = {
+    0x00, 0x00, 0x2C, 0x07,
+    0x00, 0x00,
+    0x10, 0x00,
+    0x80, 0x00,
+    0x00, 0x80,
+    0x5F, 0x9B, 0x34, 0xFB,
+};
+#endif // CONFIG_USE_BME280
+
+#ifdef CONFIG_USE_MPU6050
+#define BT_UUID_GYRO        BT_UUID_DECLARE_16(0x2c08)
+custom_uuid[] = {
+    0x00, 0x00, 0x2C, 0x08,
+    0x00, 0x00,
+    0x10, 0x00,
+    0x80, 0x00,
+    0x00, 0x80,
+    0x5F, 0x9B, 0x34, 0xFB,
+};
+#endif // CONFIG_USE_MPU6050
+
+#endif // SENSORS_UUID_H
